Refactor pusher to allow multiple implementations
Hides the pusher/Publisher details behind an interface and adds a
registry and a feature flag so that multiple implementations can
coexist.
